diy_brain/requirements.txt
psycopg2-binary
django-crispy-forms
django-bootstrap4
django-common-helpers
djongo
jsonfield
